🛤️🐚📂🧰: move GitHub API calls into `lively.git`
